requirements/local.txt
-r base.txt

ipython==1.1.0
django-extensions==1.2.5

